+++ realm/realm-annotations-processor/src/test/resources/io/realm/NullTypesRealmProxy.java
@@ -154,330 +154,330 @@ public class NullTypesRealmProxy extends NullTypes
 
     @SuppressWarnings("cast")
     public String realmGet$fieldStringNotNull() {
-        realm.checkIfValid();
-        return (java.lang.String) row.getString(columnInfo.fieldStringNotNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (java.lang.String) ((RealmObject) this).row.getString(columnInfo.fieldStringNotNullIndex);
     }
 
     public void realmSet$fieldStringNotNull(String value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
             throw new IllegalArgumentException("Trying to set non-nullable field fieldStringNotNull to null.");
         }
-        row.setString(columnInfo.fieldStringNotNullIndex, value);
+        ((RealmObject) this).row.setString(columnInfo.fieldStringNotNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public String realmGet$fieldStringNull() {
-        realm.checkIfValid();
-        return (java.lang.String) row.getString(columnInfo.fieldStringNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (java.lang.String) ((RealmObject) this).row.getString(columnInfo.fieldStringNullIndex);
     }
 
     public void realmSet$fieldStringNull(String value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.setNull(columnInfo.fieldStringNullIndex);
+            ((RealmObject) this).row.setNull(columnInfo.fieldStringNullIndex);
             return;
         }
-        row.setString(columnInfo.fieldStringNullIndex, value);
+        ((RealmObject) this).row.setString(columnInfo.fieldStringNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Boolean realmGet$fieldBooleanNotNull() {
-        realm.checkIfValid();
-        return (boolean) row.getBoolean(columnInfo.fieldBooleanNotNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (boolean) ((RealmObject) this).row.getBoolean(columnInfo.fieldBooleanNotNullIndex);
     }
 
     public void realmSet$fieldBooleanNotNull(Boolean value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
             throw new IllegalArgumentException("Trying to set non-nullable field fieldBooleanNotNull to null.");
         }
-        row.setBoolean(columnInfo.fieldBooleanNotNullIndex, value);
+        ((RealmObject) this).row.setBoolean(columnInfo.fieldBooleanNotNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Boolean realmGet$fieldBooleanNull() {
-        realm.checkIfValid();
-        if (row.isNull(columnInfo.fieldBooleanNullIndex)) {
+        ((RealmObject) this).realm.checkIfValid();
+        if (((RealmObject) this).row.isNull(columnInfo.fieldBooleanNullIndex)) {
             return null;
         }
-        return (boolean) row.getBoolean(columnInfo.fieldBooleanNullIndex);
+        return (boolean) ((RealmObject) this).row.getBoolean(columnInfo.fieldBooleanNullIndex);
     }
 
     public void realmSet$fieldBooleanNull(Boolean value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.setNull(columnInfo.fieldBooleanNullIndex);
+            ((RealmObject) this).row.setNull(columnInfo.fieldBooleanNullIndex);
             return;
         }
-        row.setBoolean(columnInfo.fieldBooleanNullIndex, value);
+        ((RealmObject) this).row.setBoolean(columnInfo.fieldBooleanNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public byte[] realmGet$fieldBytesNotNull() {
-        realm.checkIfValid();
-        return (byte[]) row.getBinaryByteArray(columnInfo.fieldBytesNotNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (byte[]) ((RealmObject) this).row.getBinaryByteArray(columnInfo.fieldBytesNotNullIndex);
     }
 
     public void realmSet$fieldBytesNotNull(byte[] value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
             throw new IllegalArgumentException("Trying to set non-nullable field fieldBytesNotNull to null.");
         }
-        row.setBinaryByteArray(columnInfo.fieldBytesNotNullIndex, value);
+        ((RealmObject) this).row.setBinaryByteArray(columnInfo.fieldBytesNotNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public byte[] realmGet$fieldBytesNull() {
-        realm.checkIfValid();
-        return (byte[]) row.getBinaryByteArray(columnInfo.fieldBytesNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (byte[]) ((RealmObject) this).row.getBinaryByteArray(columnInfo.fieldBytesNullIndex);
     }
 
     public void realmSet$fieldBytesNull(byte[] value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.setNull(columnInfo.fieldBytesNullIndex);
+            ((RealmObject) this).row.setNull(columnInfo.fieldBytesNullIndex);
             return;
         }
-        row.setBinaryByteArray(columnInfo.fieldBytesNullIndex, value);
+        ((RealmObject) this).row.setBinaryByteArray(columnInfo.fieldBytesNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Byte realmGet$fieldByteNotNull() {
-        realm.checkIfValid();
-        return (byte) row.getLong(columnInfo.fieldByteNotNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (byte) ((RealmObject) this).row.getLong(columnInfo.fieldByteNotNullIndex);
     }
 
     public void realmSet$fieldByteNotNull(Byte value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
             throw new IllegalArgumentException("Trying to set non-nullable field fieldByteNotNull to null.");
         }
-        row.setLong(columnInfo.fieldByteNotNullIndex, value);
+        ((RealmObject) this).row.setLong(columnInfo.fieldByteNotNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Byte realmGet$fieldByteNull() {
-        realm.checkIfValid();
-        if (row.isNull(columnInfo.fieldByteNullIndex)) {
+        ((RealmObject) this).realm.checkIfValid();
+        if (((RealmObject) this).row.isNull(columnInfo.fieldByteNullIndex)) {
             return null;
         }
-        return (byte) row.getLong(columnInfo.fieldByteNullIndex);
+        return (byte) ((RealmObject) this).row.getLong(columnInfo.fieldByteNullIndex);
     }
 
     public void realmSet$fieldByteNull(Byte value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.setNull(columnInfo.fieldByteNullIndex);
+            ((RealmObject) this).row.setNull(columnInfo.fieldByteNullIndex);
             return;
         }
-        row.setLong(columnInfo.fieldByteNullIndex, value);
+        ((RealmObject) this).row.setLong(columnInfo.fieldByteNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Short realmGet$fieldShortNotNull() {
-        realm.checkIfValid();
-        return (short) row.getLong(columnInfo.fieldShortNotNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (short) ((RealmObject) this).row.getLong(columnInfo.fieldShortNotNullIndex);
     }
 
     public void realmSet$fieldShortNotNull(Short value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
             throw new IllegalArgumentException("Trying to set non-nullable field fieldShortNotNull to null.");
         }
-        row.setLong(columnInfo.fieldShortNotNullIndex, value);
+        ((RealmObject) this).row.setLong(columnInfo.fieldShortNotNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Short realmGet$fieldShortNull() {
-        realm.checkIfValid();
-        if (row.isNull(columnInfo.fieldShortNullIndex)) {
+        ((RealmObject) this).realm.checkIfValid();
+        if (((RealmObject) this).row.isNull(columnInfo.fieldShortNullIndex)) {
             return null;
         }
-        return (short) row.getLong(columnInfo.fieldShortNullIndex);
+        return (short) ((RealmObject) this).row.getLong(columnInfo.fieldShortNullIndex);
     }
 
     public void realmSet$fieldShortNull(Short value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.setNull(columnInfo.fieldShortNullIndex);
+            ((RealmObject) this).row.setNull(columnInfo.fieldShortNullIndex);
             return;
         }
-        row.setLong(columnInfo.fieldShortNullIndex, value);
+        ((RealmObject) this).row.setLong(columnInfo.fieldShortNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Integer realmGet$fieldIntegerNotNull() {
-        realm.checkIfValid();
-        return (int) row.getLong(columnInfo.fieldIntegerNotNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (int) ((RealmObject) this).row.getLong(columnInfo.fieldIntegerNotNullIndex);
     }
 
     public void realmSet$fieldIntegerNotNull(Integer value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
             throw new IllegalArgumentException("Trying to set non-nullable field fieldIntegerNotNull to null.");
         }
-        row.setLong(columnInfo.fieldIntegerNotNullIndex, value);
+        ((RealmObject) this).row.setLong(columnInfo.fieldIntegerNotNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Integer realmGet$fieldIntegerNull() {
-        realm.checkIfValid();
-        if (row.isNull(columnInfo.fieldIntegerNullIndex)) {
+        ((RealmObject) this).realm.checkIfValid();
+        if (((RealmObject) this).row.isNull(columnInfo.fieldIntegerNullIndex)) {
             return null;
         }
-        return (int) row.getLong(columnInfo.fieldIntegerNullIndex);
+        return (int) ((RealmObject) this).row.getLong(columnInfo.fieldIntegerNullIndex);
     }
 
     public void realmSet$fieldIntegerNull(Integer value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.setNull(columnInfo.fieldIntegerNullIndex);
+            ((RealmObject) this).row.setNull(columnInfo.fieldIntegerNullIndex);
             return;
         }
-        row.setLong(columnInfo.fieldIntegerNullIndex, value);
+        ((RealmObject) this).row.setLong(columnInfo.fieldIntegerNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Long realmGet$fieldLongNotNull() {
-        realm.checkIfValid();
-        return (long) row.getLong(columnInfo.fieldLongNotNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (long) ((RealmObject) this).row.getLong(columnInfo.fieldLongNotNullIndex);
     }
 
     public void realmSet$fieldLongNotNull(Long value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
             throw new IllegalArgumentException("Trying to set non-nullable field fieldLongNotNull to null.");
         }
-        row.setLong(columnInfo.fieldLongNotNullIndex, value);
+        ((RealmObject) this).row.setLong(columnInfo.fieldLongNotNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Long realmGet$fieldLongNull() {
-        realm.checkIfValid();
-        if (row.isNull(columnInfo.fieldLongNullIndex)) {
+        ((RealmObject) this).realm.checkIfValid();
+        if (((RealmObject) this).row.isNull(columnInfo.fieldLongNullIndex)) {
             return null;
         }
-        return (long) row.getLong(columnInfo.fieldLongNullIndex);
+        return (long) ((RealmObject) this).row.getLong(columnInfo.fieldLongNullIndex);
     }
 
     public void realmSet$fieldLongNull(Long value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.setNull(columnInfo.fieldLongNullIndex);
+            ((RealmObject) this).row.setNull(columnInfo.fieldLongNullIndex);
             return;
         }
-        row.setLong(columnInfo.fieldLongNullIndex, value);
+        ((RealmObject) this).row.setLong(columnInfo.fieldLongNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Float realmGet$fieldFloatNotNull() {
-        realm.checkIfValid();
-        return (float) row.getFloat(columnInfo.fieldFloatNotNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (float) ((RealmObject) this).row.getFloat(columnInfo.fieldFloatNotNullIndex);
     }
 
     public void realmSet$fieldFloatNotNull(Float value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
             throw new IllegalArgumentException("Trying to set non-nullable field fieldFloatNotNull to null.");
         }
-        row.setFloat(columnInfo.fieldFloatNotNullIndex, value);
+        ((RealmObject) this).row.setFloat(columnInfo.fieldFloatNotNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Float realmGet$fieldFloatNull() {
-        realm.checkIfValid();
-        if (row.isNull(columnInfo.fieldFloatNullIndex)) {
+        ((RealmObject) this).realm.checkIfValid();
+        if (((RealmObject) this).row.isNull(columnInfo.fieldFloatNullIndex)) {
             return null;
         }
-        return (float) row.getFloat(columnInfo.fieldFloatNullIndex);
+        return (float) ((RealmObject) this).row.getFloat(columnInfo.fieldFloatNullIndex);
     }
 
     public void realmSet$fieldFloatNull(Float value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.setNull(columnInfo.fieldFloatNullIndex);
+            ((RealmObject) this).row.setNull(columnInfo.fieldFloatNullIndex);
             return;
         }
-        row.setFloat(columnInfo.fieldFloatNullIndex, value);
+        ((RealmObject) this).row.setFloat(columnInfo.fieldFloatNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Double realmGet$fieldDoubleNotNull() {
-        realm.checkIfValid();
-        return (double) row.getDouble(columnInfo.fieldDoubleNotNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (double) ((RealmObject) this).row.getDouble(columnInfo.fieldDoubleNotNullIndex);
     }
 
     public void realmSet$fieldDoubleNotNull(Double value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
             throw new IllegalArgumentException("Trying to set non-nullable field fieldDoubleNotNull to null.");
         }
-        row.setDouble(columnInfo.fieldDoubleNotNullIndex, value);
+        ((RealmObject) this).row.setDouble(columnInfo.fieldDoubleNotNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Double realmGet$fieldDoubleNull() {
-        realm.checkIfValid();
-        if (row.isNull(columnInfo.fieldDoubleNullIndex)) {
+        ((RealmObject) this).realm.checkIfValid();
+        if (((RealmObject) this).row.isNull(columnInfo.fieldDoubleNullIndex)) {
             return null;
         }
-        return (double) row.getDouble(columnInfo.fieldDoubleNullIndex);
+        return (double) ((RealmObject) this).row.getDouble(columnInfo.fieldDoubleNullIndex);
     }
 
     public void realmSet$fieldDoubleNull(Double value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.setNull(columnInfo.fieldDoubleNullIndex);
+            ((RealmObject) this).row.setNull(columnInfo.fieldDoubleNullIndex);
             return;
         }
-        row.setDouble(columnInfo.fieldDoubleNullIndex, value);
+        ((RealmObject) this).row.setDouble(columnInfo.fieldDoubleNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Date realmGet$fieldDateNotNull() {
-        realm.checkIfValid();
-        return (java.util.Date) row.getDate(columnInfo.fieldDateNotNullIndex);
+        ((RealmObject) this).realm.checkIfValid();
+        return (java.util.Date) ((RealmObject) this).row.getDate(columnInfo.fieldDateNotNullIndex);
     }
 
     public void realmSet$fieldDateNotNull(Date value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
             throw new IllegalArgumentException("Trying to set non-nullable field fieldDateNotNull to null.");
         }
-        row.setDate(columnInfo.fieldDateNotNullIndex, value);
+        ((RealmObject) this).row.setDate(columnInfo.fieldDateNotNullIndex, value);
     }
 
     @SuppressWarnings("cast")
     public Date realmGet$fieldDateNull() {
-        realm.checkIfValid();
-        if (row.isNull(columnInfo.fieldDateNullIndex)) {
+        ((RealmObject) this).realm.checkIfValid();
+        if (((RealmObject) this).row.isNull(columnInfo.fieldDateNullIndex)) {
             return null;
         }
-        return (java.util.Date) row.getDate(columnInfo.fieldDateNullIndex);
+        return (java.util.Date) ((RealmObject) this).row.getDate(columnInfo.fieldDateNullIndex);
     }
 
     public void realmSet$fieldDateNull(Date value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.setNull(columnInfo.fieldDateNullIndex);
+            ((RealmObject) this).row.setNull(columnInfo.fieldDateNullIndex);
             return;
         }
-        row.setDate(columnInfo.fieldDateNullIndex, value);
+        ((RealmObject) this).row.setDate(columnInfo.fieldDateNullIndex, value);
     }
 
     public NullTypes realmGet$fieldObjectNull() {
-        realm.checkIfValid();
-        if (row.isNullLink(columnInfo.fieldObjectNullIndex)) {
+        ((RealmObject) this).realm.checkIfValid();
+        if (((RealmObject) this).row.isNullLink(columnInfo.fieldObjectNullIndex)) {
             return null;
         }
-        return realm.get(some.test.NullTypes.class, row.getLink(columnInfo.fieldObjectNullIndex));
+        return ((RealmObject) this).realm.get(some.test.NullTypes.class, ((RealmObject) this).row.getLink(columnInfo.fieldObjectNullIndex));
     }
 
     public void realmSet$fieldObjectNull(NullTypes value) {
-        realm.checkIfValid();
+        ((RealmObject) this).realm.checkIfValid();
         if (value == null) {
-            row.nullifyLink(columnInfo.fieldObjectNullIndex);
+            ((RealmObject) this).row.nullifyLink(columnInfo.fieldObjectNullIndex);
             return;
         }
         if (!value.isValid()) {
@@ -486,7 +486,7 @@ public class NullTypesRealmProxy extends NullTypes
         if (value.realm != this.realm) {
             throw new IllegalArgumentException("'value' belongs to a different Realm.");
         }
-        row.setLink(columnInfo.fieldObjectNullIndex, value.row.getIndex());
+        ((RealmObject) this).row.setLink(columnInfo.fieldObjectNullIndex, ((RealmObject) value).row.getIndex());
     }
 
     public static Table initTable(ImplicitTransaction transaction) {
@@ -744,7 +744,7 @@ public class NullTypesRealmProxy extends NullTypes
 
     @SuppressWarnings("cast")
     public static NullTypes createOrUpdateUsingJsonObject(Realm realm, JSONObject json, boolean update)
-        throws JSONException {
+            throws JSONException {
         NullTypes obj = realm.createObject(NullTypes.class);
         if (json.has("fieldStringNotNull")) {
             if (json.isNull("fieldStringNotNull")) {
@@ -909,7 +909,7 @@ public class NullTypesRealmProxy extends NullTypes
 
     @SuppressWarnings("cast")
     public static NullTypes createUsingJsonStream(Realm realm, JsonReader reader)
-        throws IOException {
+            throws IOException {
         NullTypes obj = realm.createObject(NullTypes.class);
         reader.beginObject();
         while (reader.hasNext()) {
@@ -1081,7 +1081,7 @@ public class NullTypesRealmProxy extends NullTypes
     }
 
     public static NullTypes copyOrUpdate(Realm realm, NullTypes object, boolean update, Map<RealmObject,RealmObjectProxy> cache) {
-        if (object.realm != null && object.realm.getPath().equals(realm.getPath())) {
+        if (((RealmObject) object).realm != null && ((RealmObject) object).realm.getPath().equals(realm.getPath())) {
             return object;
         }
         return copy(realm, object, update, cache);
@@ -1090,26 +1090,26 @@ public class NullTypesRealmProxy extends NullTypes
     public static NullTypes copy(Realm realm, NullTypes newObject, boolean update, Map<RealmObject,RealmObjectProxy> cache) {
         NullTypes realmObject = realm.createObject(NullTypes.class);
         cache.put(newObject, (RealmObjectProxy) realmObject);
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldStringNotNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldStringNotNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldStringNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldStringNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldBooleanNotNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldBooleanNotNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldBooleanNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldBooleanNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldBytesNotNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldBytesNotNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldBytesNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldBytesNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldByteNotNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldByteNotNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldByteNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldByteNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldShortNotNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldShortNotNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldShortNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldShortNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldIntegerNotNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldIntegerNotNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldIntegerNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldIntegerNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldLongNotNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldLongNotNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldLongNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldLongNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldFloatNotNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldFloatNotNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldFloatNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldFloatNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldDoubleNotNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldDoubleNotNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldDoubleNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldDoubleNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldDateNotNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldDateNotNull());
-        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldDateNull(((NullTypesRealmProxyInterface)newObject).realmGet$fieldDateNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldStringNotNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldStringNotNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldStringNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldStringNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldBooleanNotNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldBooleanNotNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldBooleanNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldBooleanNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldBytesNotNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldBytesNotNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldBytesNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldBytesNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldByteNotNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldByteNotNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldByteNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldByteNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldShortNotNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldShortNotNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldShortNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldShortNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldIntegerNotNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldIntegerNotNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldIntegerNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldIntegerNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldLongNotNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldLongNotNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldLongNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldLongNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldFloatNotNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldFloatNotNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldFloatNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldFloatNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldDoubleNotNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldDoubleNotNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldDoubleNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldDoubleNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldDateNotNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldDateNotNull());
+        ((NullTypesRealmProxyInterface) realmObject).realmSet$fieldDateNull(((NullTypesRealmProxyInterface) newObject).realmGet$fieldDateNull());
 
         some.test.NullTypes fieldObjectNullObj = ((NullTypesRealmProxyInterface) newObject).realmGet$fieldObjectNull();
         if (fieldObjectNullObj != null) {
@@ -1264,9 +1264,9 @@ public class NullTypesRealmProxy extends NullTypes
 
     @Override
     public int hashCode() {
-        String realmName = realm.getPath();
-        String tableName = row.getTable().getName();
-        long rowIndex = row.getIndex();
+        String realmName = ((RealmObject) this).realm.getPath();
+        String tableName = ((RealmObject) this).row.getTable().getName();
+        long rowIndex = ((RealmObject) this).row.getIndex();
 
         int result = 17;
         result = 31 * result + ((realmName != null) ? realmName.hashCode() : 0);
@@ -1281,15 +1281,15 @@ public class NullTypesRealmProxy extends NullTypes
         if (o == null || getClass() != o.getClass()) return false;
         NullTypesRealmProxy aNullTypes = (NullTypesRealmProxy)o;
 
-        String path = realm.getPath();
-        String otherPath = aNullTypes.realm.getPath();
+        String path = ((RealmObject) this).realm.getPath();
+        String otherPath = ((RealmObject) aNullTypes).realm.getPath();
         if (path != null ? !path.equals(otherPath) : otherPath != null) return false;;
 
-        String tableName = row.getTable().getName();
-        String otherTableName = aNullTypes.row.getTable().getName();
+        String tableName = ((RealmObject) this).row.getTable().getName();
+        String otherTableName = ((RealmObject) aNullTypes).row.getTable().getName();
         if (tableName != null ? !tableName.equals(otherTableName) : otherTableName != null) return false;
 
-        if (row.getIndex() != aNullTypes.row.getIndex()) return false;
+        if (((RealmObject) this).row.getIndex() != ((RealmObject) aNullTypes).row.getIndex()) return false;
 
         return true;
     }